Fix clippy warnings in Rust workspace; add TODO(exarp) to allow(dead_code) stubs
- api: simplify booleans, remove redundant into_iter(), allow(dead_code) + TODO(exarp) for FMP fundamentals
- ib_adapter, quant, risk: allow(dead_code)/clippy fixes as needed
- tui_service: event routing allow(dead_code) + TODO(T-1773357423959019000); SnapshotSource Rest + TODO(T-1773357423959019000); doc/println/config/app fixes
- backend_service: clone_on_copy fix in collection_aggregation


Former-commit-id: a505f1bdad9c0f5c7907755937ef491d1a3f17d1

+++ agents/backend/services/tui_service/src/config.rs
@@ -101,11 +101,11 @@ impl TuiConfig {
         }
 
         if let Some(tick_ms) = shared.tui.refresh_rate_ms {
-            self.tick_ms = tick_ms.max(1) as u64;
+            self.tick_ms = tick_ms.max(1);
         }
 
         if let Some(rest_poll_ms) = shared.tui.update_interval_ms {
-            self.rest_poll_ms = rest_poll_ms.max(1) as u64;
+            self.rest_poll_ms = rest_poll_ms.max(1);
         }
 
         if let Some(symbols) = shared.strategy.symbols.as_ref() {
